Refactoring with plugin-audio.cpp
